Refactor discount code imports to use new paths and add discount-related utility functions
