+++ apps/web/app/(ee)/api/cron/partners/deactivate/route.ts
@@ -1,6 +1,6 @@
-import { deleteDiscountCodes } from "@/lib/api/discounts/delete-discount-code";
 import { linkCache } from "@/lib/api/links/cache";
 import { withCron } from "@/lib/cron/with-cron";
+import { deleteDiscountCodes } from "@/lib/discounts/delete-discount-code";
 import { sendBatchEmail } from "@dub/email";
 import PartnerDeactivated from "@dub/email/templates/partner-deactivated";
 import { prisma } from "@dub/prisma";
