+++ packages/prisma/schema/discount.prisma
@@ -1,15 +1,21 @@
+enum DiscountProvider {
+  stripe
+  shopify
+}
+
 model Discount {
-  id                     String          @id @default(cuid())
+  id                     String           @id @default(cuid())
   programId              String
-  amount                 Int             @default(0)
-  type                   RewardStructure @default(percentage)
+  amount                 Int              @default(0)
+  type                   RewardStructure  @default(percentage)
   maxDuration            Int? // in months (0 -> one-time purchase, 1, 3, 6, 12, 18, 24, 36, 48 -> months, null -> lifetime)
   description            String?
   couponId               String?
   couponTestId           String?
   autoProvisionEnabledAt DateTime? // Auto create discount codes for partners when they join the group
-  createdAt              DateTime        @default(now())
-  updatedAt              DateTime        @updatedAt
+  provider               DiscountProvider @default(stripe)
+  createdAt              DateTime         @default(now())
+  updatedAt              DateTime         @updatedAt
 
   programEnrollments ProgramEnrollment[]
   program            Program             @relation("ProgramDiscounts", fields: [programId], references: [id])
